+++ libnavui-maps/src/test/java/com/mapbox/navigation/ui/maps/building/BuildingProcessorTest.kt
@@ -2,12 +2,21 @@ package com.mapbox.navigation.ui.maps.building
 
 import com.mapbox.api.directions.v5.models.DirectionsRoute
 import com.mapbox.api.directions.v5.models.RouteOptions
+import com.mapbox.bindgen.ExpectedFactory
 import com.mapbox.geojson.Point
+import com.mapbox.maps.MapboxMap
+import com.mapbox.maps.QueryFeaturesCallback
+import com.mapbox.maps.RenderedQueryOptions
+import com.mapbox.maps.ScreenCoordinate
 import com.mapbox.navigation.base.trip.model.RouteLegProgress
 import com.mapbox.navigation.base.trip.model.RouteProgress
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.slot
+import io.mockk.verify
+import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
 import org.junit.Test
 
 class BuildingProcessorTest {
@@ -204,5 +213,31 @@ class BuildingProcessorTest {
         val result = BuildingProcessor.queryBuildingOnFinalDestination(mockAction)
 
         assertEquals(result.point, mockFinalForCoordinates)
+    }
+
+    @Test
+    fun `queryBuilding will query correct building layers`() = runBlockingTest {
+        val callbackSlot = slot<QueryFeaturesCallback>()
+        val optionsSlot = slot<RenderedQueryOptions>()
+        val screenCoordinate = mockk<ScreenCoordinate>()
+        val mockMap = mockk<MapboxMap>(relaxed = true) {
+            every { pixelForCoordinate(any()) } returns screenCoordinate
+            every {
+                queryRenderedFeatures(any<ScreenCoordinate>(), any(), capture(callbackSlot))
+            } answers {
+                callbackSlot.captured.run(ExpectedFactory.createValue(listOf()))
+            }
+        }
+        val action = BuildingAction.QueryBuilding(
+            Point.fromLngLat(-122.4145, 37.7653),
+            mockMap
+        )
+
+        BuildingProcessor.queryBuilding(action)
+
+        verify { mockMap.queryRenderedFeatures(screenCoordinate, capture(optionsSlot), any()) }
+        assertEquals(2, optionsSlot.captured.layerIds!!.size)
+        assertTrue(optionsSlot.captured.layerIds!!.contains("building"))
+        assertTrue(optionsSlot.captured.layerIds!!.contains("building-extrusion"))
     }
 }